requirements.txt
streamlit
sentence-transformers
transformers
torch
faiss-cpu
pypdf
scikit-learn
numpy
tqdm
